Create LICENSE
Adding MIT License, as requested in the Issues. I was the original and sole author a long time ago and it was always an MIT license, so let's keep that up. Feel free to add more names to the copyright holders in case other people made changes in the meantime.